--- src/Credits.tsx
@@ -1,10 +1,0 @@
-//TODO: add file to refactor and separe the code
-import React from 'react';
-
-export const Results: React.FC = () => {
-  return (
-    <div>
-      <p>Credits</p>
-    </div>
-  );
-};
