--- src/Results.tsx
@@ -1,72 +1,0 @@
-//TODO: add file to refactor and separe the code
-import React from 'react';
-
-export const Results: React.FC = () => {
-  return (
-    <>
-      <p>Hello</p>
-      {/* <Container>
-        <Text fontSize='lg' as='b'>
-          Résultats:
-        </Text>
-        {!isFirstSemester && (
-          <Container m={1}>
-            <Text fontSize='md' as='b'>
-              Au début du semestre
-            </Text>
-            <Text>
-              Crédits:&nbsp;
-              {form?.currentCreditsEarned}
-            </Text>
-            <Text>
-              Notes pointage:&nbsp;
-              {calculateGradePoints(form?.currentCreditsEarned, form?.currentGPA)}
-            </Text>
-            <Text>
-              Moyenne cumulative globale sur 4,30:&nbsp;
-              {form?.currentGPA}
-            </Text>
-          </Container>
-        )}
-
-        <Container m={1}>
-          <Text fontSize='md' as='b'>
-            Calcul estimé pour ce semestre
-          </Text>
-          <Text>
-            Crédits:&nbsp;
-            {accumulatedCredits(form)}
-          </Text>
-          <Text>
-            Notes pointage:&nbsp;
-            {semesterGradePoints(form)}
-          </Text>
-          <Text>
-            Moyenne cumulative globale sur 4,30:&nbsp;
-            {semesterCalculateGPA(form)}
-          </Text>
-        </Container>
-
-        {!isFirstSemester && (
-          <Container m={1}>
-            <Text fontSize='md' as='b'>
-              Calcul estimé - total
-            </Text>
-            <Text>
-              Total crédits:&nbsp;
-              {totalCredits(form)}
-            </Text>
-            <Text>
-              Total notes pointage: &nbsp;
-              {finalPoints(form)}
-            </Text>
-            <Text>
-              Résultat moyenne cumulative globale sur 4,30:&nbsp;
-              {finalGPA(form)}
-            </Text>
-          </Container>
-        )} */}
-      {/* </Container> */}
-    </>
-  );
-}
